requirements.txt
APScheduler==3.6.3
pytz==2023.3
fastapi==0.89.0
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.21.0
requests==2.25.0
SQLAlchemy==1.3.20
uvicorn==0.13.4
python-multipart==0.0.6
